fix key types of event_log.

+++ backend/backend.sql
@@ -191,7 +191,7 @@ INSERT INTO DOUBUTSU.event_log_seq VALUES (0);
 -- ** 温泉のはじまり
 --  event_type == 0
 CREATE TABLE DOUBUTSU.event_start
-  ( event_log_id INTEGER NOT NULL
+  ( event_log_id BIGINT NOT NULL
   , onsen_id INTEGER NOT NULL
 
   , PRIMARY KEY (event_log_id)
@@ -201,7 +201,7 @@ CREATE TABLE DOUBUTSU.event_start
 --  event_type == 1
 --  event_data == id
 CREATE TABLE DOUBUTSU.event_action
-  ( event_log_id INTEGER NOT NULL
+  ( event_log_id BIGINT NOT NULL
   , onsen_id INTEGER NOT NULL
   , local_slot_number INTEGER NOT NULL
   , doubutsu_id INTEGER NOT NULL
@@ -214,7 +214,7 @@ CREATE TABLE DOUBUTSU.event_action
 --  event_type == 2
 --  event_data == id
 CREATE TABLE DOUBUTSU.event_item
-  ( event_log_id INTEGER NOT NULL
+  ( event_log_id BIGINT NOT NULL
   , item_id INTEGER NOT NULL
   , got_reason_type INTEGER NOT NULL
   , got_reason_data INTEGER NOT NULL
@@ -226,7 +226,7 @@ CREATE TABLE DOUBUTSU.event_item
 --  event_type == 3
 --  event_data == id
 CREATE TABLE DOUBUTSU.event_growth
-  ( event_log_id INTEGER NOT NULL
+  ( event_log_id BIGINT NOT NULL
   , onsen_id INTEGER NOT NULL
   , local_slot_number INTEGER NOT NULL
   , item_id INTEGER NOT NULL
@@ -239,7 +239,7 @@ CREATE TABLE DOUBUTSU.event_growth
 --  event_type == 4
 --  event_data == id
 CREATE TABLE DOUBUTSU.event_mission
-  ( event_log_id INTEGER NOT NULL
+  ( event_log_id BIGINT NOT NULL
   , onsen_level INTEGER NOT NULL
   , mission_status INTEGER NOT NULL -- 遷移後状態
 
@@ -249,14 +249,14 @@ CREATE TABLE DOUBUTSU.event_mission
 -- ** 時間が経つ
 --  event_type == 5
 CREATE TABLE DOUBUTSU.event_time
-  ( event_log_id INTEGER NOT NULL
+  ( event_log_id BIGINT NOT NULL
   , time_type INTEGER NOT NULL -- 朝: 0 , 昼: 1 , 夕方: 2 , 夜: 3
   );
 
 -- ** 天気が変わる
 --  event_type == 6
 CREATE TABLE DOUBUTSU.event_whether
-  ( event_log_id INTEGER NOT NULL
+  ( event_log_id BIGINT NOT NULL
   , whether INTEGER NOT NULL -- 天気の状態のID
   );
